// types/user.d.ts
export type AssetFinance = {
  purchaseCost: number;
  depreciationValue: number;
  maintenanceCost: number;
  resaleValue: number;
  purchaseDate: string;
  warranty: string;
  supplierName: string;
  supplierContact: string;
  purchaseReceipt: string;
} | null;

// Assets
export type Asset = {
  id: string;
  name: string;
  serialNumber: string;
  usefulLife: number;
  imageUrl: string;
  fiscalYearId: string;
  assetFinance?: AssetFinance;
  category: CommonType;
  branch: CommonType;
  department: CommonType;
  fiscalYear: CommonType;
  categoryId?: ICategory;
  client_BranchId?: IBranch;
  departmentId?: IDepartment;
  depreciation?: Depreciation;
  approvalStatus: "pending" | "approved" | "rejected" | string;
  approvalMessage?: string;
};

interface Depreciation {
  totalDepreciation: number;
  remainingAmountAfterDepreciation: number;
  yearsPassed: number;
}

interface CommonType {
  id: string;
  name: string;
  code?: string;
  state?: string;
  district?: string;
  addressLine?: string;
  statdate?: string;
  enddata?: string;
}